+++ tests/RequestHandler/OrderRequestHandlerTest.php
@@ -61,7 +61,7 @@ class OrderRequestHandlerTest extends TestCase
         $this->assertEquals(1, $data['paging']['page']);
         $this->assertEquals(1, $data['paging']['totalPages']);
         $this->assertEquals(1, $data['paging']['totalCount']);
-        $body = '{"paging":{"page":1,"totalCount":1,"totalPages":1},"orders":[{"order_id":"1234","order_number":"456","currency_code":"DE","nick_name":"GirlWhoCanFly","ship_cost":4.9,"invoice_address":{"firstname":"Kara","lastname":"Zor-El","street":"Argo Street","housenumber":"1022","address2":"Window","postcode":"90012","city":"National City","country_code":"US","company":"D.E.O.","state":"CA"},"delivery_address":{"firstname":"Kara","lastname":"Zor-El","street":"Argo Street","housenumber":"1022","address2":"Window","postcode":"90012","city":"National City","country_code":"US","company":"D.E.O.","state":"CA"},"order_date":"2019-01-01T20:00:15+0000","email":"secret@deo.tld","phone1":"0123456789","pay_date":"2019-01-01T23:00:15+0000","ship_date":"2019-01-02T02:00:15+0000","payment_method":1,"order_status_id":2,"seller_comment":"Psst","shippingprofile_id":"super-fast","vat_id":"DE-123456","payment_transaction_id":"123444"}]}';
+        $body = '{"paging":{"page":1,"totalCount":1,"totalPages":1},"orders":[{"order_id":"1234","order_number":"456","currency_code":"EUR","delivery_source_country_code":"DE","nick_name":"GirlWhoCanFly","ship_cost":4.9,"invoice_address":{"firstname":"Kara","lastname":"Zor-El","street":"Argo Street","housenumber":"1022","address2":"Window","postcode":"90012","city":"National City","country_code":"US","company":"D.E.O.","state":"CA"},"delivery_address":{"firstname":"Kara","lastname":"Zor-El","street":"Argo Street","housenumber":"1022","address2":"Window","postcode":"90012","city":"National City","country_code":"US","company":"D.E.O.","state":"CA"},"order_date":"2019-01-01T20:00:15+0000","email":"secret@deo.tld","phone1":"0123456789","pay_date":"2019-01-01T23:00:15+0000","ship_date":"2019-01-02T02:00:15+0000","payment_method":1,"order_status_id":2,"seller_comment":"Psst","shippingprofile_id":"super-fast","vat_id":"DE-123456","payment_transaction_id":"123444"}]}';
         $this->assertEquals($body, $response->getBody());
     }
 
@@ -104,10 +104,11 @@ class OrderRequestHandlerTest extends TestCase
         $handler = new OrderRequestHandler($repo);
 
         $request = new Request();
-        $uri = new Uri('http://localhost/?Action=AckOrder&OrderId=1');
-        $req = $request->withUri($uri);
+        $uri = new Uri('http://localhost/?Action=AckOrder');
+        $req = $request->withUri($uri)
+                       ->withBody(new MemoryStream(json_encode(['OrderId' => '1'])));
 
-        $response = $handler->handle($req, ['Action' => 'AckOrder', 'OrderId' => '1']);
+        $response = $handler->handle($req, ['Action' => 'AckOrder']);
 
         $this->assertEquals(404, $response->getStatusCode());
         $this->assertEquals('Die Bestellung wurde nicht gefunden', (string)$response->getBody());
@@ -122,9 +123,11 @@ class OrderRequestHandlerTest extends TestCase
 
         $request = new Request();
         $uri = new Uri('http://localhost/?Action=AckOrder&OrderId=1');
-        $req = $request->withUri($uri);
+        $req = $request->withUri($uri)
+                       ->withBody(new MemoryStream(json_encode(['OrderId' => '1'])));
 
-        $response = $handler->handle($req, ['Action' => 'AckOrder', 'OrderId' => '1']);
+
+        $response = $handler->handle($req, ['Action' => 'AckOrder']);
 
         $this->assertEquals(500, $response->getStatusCode());
         $this->assertEquals('Unknown Error', (string)$response->getBody());
@@ -139,9 +142,11 @@ class OrderRequestHandlerTest extends TestCase
 
         $request = new Request();
         $uri = new Uri('http://localhost/?Action=AckOrder&OrderId=1');
-        $req = $request->withUri($uri);
+        $req = $request->withUri($uri)
+                       ->withBody(new MemoryStream(json_encode(['OrderId' => '1'])));
 
-        $response = $handler->handle($req, ['Action' => 'AckOrder', 'OrderId' => '1']);
+
+        $response = $handler->handle($req, ['Action' => 'AckOrder']);
 
         $this->assertEquals(200, $response->getStatusCode());
     }
@@ -210,7 +215,7 @@ class OrderRequestHandlerTest extends TestCase
         $response = $handler->handle($req, ['Action' => 'GetOrder', 'OrderId' => '1']);
 
         $this->assertEquals(200, $response->getStatusCode());
-        $bodyJson = '{"order_id":"1234","order_number":"456","currency_code":"DE","nick_name":"GirlWhoCanFly","ship_cost":4.9,"invoice_address":{"firstname":"Kara","lastname":"Zor-El","street":"Argo Street","housenumber":"1022","address2":"Window","postcode":"90012","city":"National City","country_code":"US","company":"D.E.O.","state":"CA"},"delivery_address":{"firstname":"Kara","lastname":"Zor-El","street":"Argo Street","housenumber":"1022","address2":"Window","postcode":"90012","city":"National City","country_code":"US","company":"D.E.O.","state":"CA"},"order_date":"2019-01-01T20:00:15+0000","email":"secret@deo.tld","phone1":"0123456789","pay_date":"2019-01-01T23:00:15+0000","ship_date":"2019-01-02T02:00:15+0000","payment_method":1,"order_status_id":2,"seller_comment":"Psst","shippingprofile_id":"super-fast","vat_id":"DE-123456","payment_transaction_id":"123444"}';
+        $bodyJson = '{"order_id":"1234","order_number":"456","currency_code":"EUR","delivery_source_country_code":"DE","nick_name":"GirlWhoCanFly","ship_cost":4.9,"invoice_address":{"firstname":"Kara","lastname":"Zor-El","street":"Argo Street","housenumber":"1022","address2":"Window","postcode":"90012","city":"National City","country_code":"US","company":"D.E.O.","state":"CA"},"delivery_address":{"firstname":"Kara","lastname":"Zor-El","street":"Argo Street","housenumber":"1022","address2":"Window","postcode":"90012","city":"National City","country_code":"US","company":"D.E.O.","state":"CA"},"order_date":"2019-01-01T20:00:15+0000","email":"secret@deo.tld","phone1":"0123456789","pay_date":"2019-01-01T23:00:15+0000","ship_date":"2019-01-02T02:00:15+0000","payment_method":1,"order_status_id":2,"seller_comment":"Psst","shippingprofile_id":"super-fast","vat_id":"DE-123456","payment_transaction_id":"123444"}';
         $this->assertEquals($bodyJson, (string)$response->getBody());
     }
 
@@ -318,7 +323,8 @@ class OrderRequestHandlerTest extends TestCase
         return (new Order())
             ->setOrderId('1234')
             ->setOrderNumber('456')
-            ->setCountryCode('DE')
+            ->setCurrencyCode('EUR')
+            ->setDeliverySourceCountryCode('DE')
             ->setNickName('GirlWhoCanFly')
             ->setShipCost(4.90)
             ->setInvoiceAddress($address)