+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
-## Next
+## v2.0.0
+
+Added:
+- `Order::$deliverySourceCountryCode`.
+
 Bug Fixes:
+- Renamed `Order::$countryCode` to `Order::$currencyCode` (breaking)
+- Fixed the unit tests
 - Fix method ackOrder - change url arguments to post body ([PR #1](https://github.com/billbeeio/custom-shop-php-sdk/pull/1))
 
 ## v1.0.1 (2. Sept. 2019)